requirements.txt
dask
numpy
Pillow
